change data in controller routes
